Add simple LBVH builder
This LBVH is based on the paper: "Maximizing Parallelism in the Construcion of
BVHs, Octrees and k-d Trees", by Tero Karras 2012. It's not as fast, since the
core of the algorithm is sequential (and simplified), but it should work.

The BVH dumped to disk currently stores the vertices and indices, and the BVH
stores references to the indices. This will probably create a smaller tree, but
may not be as performant as storing the primitives on the leaves themselves.