Add support for BVH4. Fix Refit bug.
Turns out that sorting the boxes breaks the correspondence between a box's index
and its TriID (when the boxes are first created, the box's index matches its
TriID). So, when refitting, we can no longer use T.TriID to access the box of
the primitive through boxes[T.TriID]. Instead, we just have to use T.box.

Also did some code refactoring and renamed boxRefIdx to TriID because this name
makes more sense