Save thresholds profiles. Make thresholds textbox multiline. Add OPT CRCs.
In order to make an automatic tool that smooths a bunch of OPTs, we need to
first save the thresholds applied when doing the manual smoothing. This commit
saves those threshold profiles in .thr files.

Each thresholds profile contains a CRC computed from the OPT vertices. So if
they change, we can avoid applying the profile later.